Removed module entry
Sorry if I missed something, but when I `npm i` this package I do not have any `dist` folder. Hence some bundlers (e.g., `parcel`) break as they respect the `module` field and do not fallback in situations when the entry is invalid.

(The option to fix this is to actually create / supply the ESM output to dist, but I did not look close enough to see what needs to be changed to enable this (again))

Thanks for the great project! :+1: